changing adder icons and note icon in toolbar.
Addressing some PR comments.

Changed styling of the adder a bit.
Changed annotatoin icon click on the toolbar to create a new annotation instead of a new page note, when text has been selected.
Added a for attr for icon labels in the adder.

getting scripts/update-icon-font from branch: annotation-card-design-refresh

comitting new woff and icomoon css

showing adder in the middle top of selection rather than mousePosition.

Revert "showing adder in the middle top of selection rather than mousePosition."

Implement new design for adder and toolbar

Implement the new design for
the Annotate and Highlight buttons
in the adder and the New Page Note button in the toolbar (https://trello.com/c/w6UIP3xw/281-update-the-toolbar-s-note-icon-tooltip-and-update-the-adder-design)
 * Update icons for adder and toolbar
 * Add labels to adder buttons
 * BEM-ify CSS styling