WIP - Convert viewFilter to a plain CommonJS module
Convert viewFilter from an Angular service to a plain CommonJS module
which exports a set of functions. This removes a bunch of Angular
boilerplate from the tests and makes it clearer that the module is not
stateful but just a set of functions.

 * Avoid exporting viewFilter.fields purely for use in tests.
   Instead allow the set of field matchers to be passed as an argument
   to the filter() function.

TODO - Expand on the rationale for this change.