WIP: Add options page to extension, allow disabling badge
This is an experimental commit to add an options pane to the Chrome
extension, with the first available option being the ability to disable
the annotation count badge.

The settings are persisted in a dictionary using the `chrome.storage`
API, and used to disable the code responsible for updating the badge
state when a tab changes.