Rename h.api to memex
This moves the code responsible for the presumed-reusable parts of the h
application, namely the code implementing the annotation API, into its
own project, called memex.

The motivation for this work was outlined previously [here][1] and
[here][2] and will be summarised in an ADR in due course.

The memex package moves into a `src/` directory in order to make it easy
to ensure we test the memex package as installed.

[1]: https://groups.google.com/a/list.hypothes.is/forum/#!topic/dev/6q7q8ccnKkE
[2]: https://groups.google.com/a/list.hypothes.is/forum/#!topic/dev/hOjn4UfU1Ww